+++ com.nxp.timedoctor.core.model/src/com/nxp/timedoctor/core/model/lines/MemCyclesSampleLine.java
@@ -54,8 +54,8 @@ public class MemCyclesSampleLine extends SampleLine {
 		for (int i = 0; i < getCount(); i++) {
 			getSample(i).val += baseVal;
 			if (getSample(i).val < val) {
-				baseVal += (double) OVERFLOW_CONSTANT;
-				getSample(i).val += (double) OVERFLOW_CONSTANT;
+				baseVal += OVERFLOW_CONSTANT;
+				getSample(i).val += OVERFLOW_CONSTANT;
 			}
 			val = getSample(i).val;
 		}
